+++ frame_forge/__init__.py
@@ -42,6 +42,10 @@ class GenerateImages:
         sub_outline_color: str,
         sub_back_color: str,
         sub_font_name: str,
+        sub_bold: int,
+        sub_italic: int,
+        sub_underline: int,
+        sub_strikeout: int,
         source_sub_title: str,
         release_sub_title: str,
     ):
@@ -75,6 +79,10 @@ class GenerateImages:
         self.sub_outline_color = sub_outline_color
         self.sub_back_color = sub_back_color
         self.sub_font_name = sub_font_name
+        self.sub_bold = sub_bold
+        self.sub_italic = sub_italic
+        self.sub_underline = sub_underline
+        self.sub_strikeout = sub_strikeout
         self.source_sub_title = source_sub_title
         self.release_sub_title = release_sub_title
 
@@ -119,7 +127,8 @@ class GenerateImages:
 
         selected_sub_style = (
             f"{self.sub_font_name},{self.sub_size},{color},{secondary_color},{outline_color},{back_color},"
-            f"1,0,0,0,100,100,0,0,1,1,0,{self.sub_alignment},10,10,10,1"
+            f"{self.sub_bold},{self.sub_italic},{self.sub_underline},{self.sub_strikeout},"
+            f"100,100,0,0,1,1,0,{self.sub_alignment},10,10,10,1"
         )
         sync_sub_base = (
             "Segoe UI,{size},&H31FF31&,&H00000000,&H00000000,&H00000000,"